feat: add ability to include non-data elements to WidgetMapper
This commit alters the way widgets are mapped onto
WidgetMapper. Instead of passing the WidgetMapper functional component
as the mapping function, it creates an arrow function that manually
maps each widget data attribute to a WidgetMapper prop. This somewhat
verbose solution is needed in order to allow WidgetMapper to receive
other props that don't come from the `widgets` data structure, such as
functions.